+++ packages/vscode-extension/lib/wrapper.js
@@ -12,7 +12,8 @@ const {
   findNodeHandle,
 } = require("react-native");
 const { storybookPreview } = require("./storybook_helper");
- 
+const { DimensionsBox } = require("./components/DimensionsBox").default;
+
 // https://github.com/facebook/react/blob/c3570b158d087eb4e3ee5748c4bd9360045c8a26/packages/react-reconciler/src/ReactWorkTags.js#L62
 const OffscreenComponentReactTag = 22;
 
@@ -26,7 +27,7 @@ let navigationHistory = new Map();
 const InternalImports = {
   get PREVIEW_APP_KEY() {
     return require("./preview").PREVIEW_APP_KEY;
-  },
+  }
 };
 
 const RNInternals = {
@@ -196,7 +197,7 @@ export function AppWrapper({ children, initialProps, fabric }) {
   const [devtoolsAgent, setDevtoolsAgent] = useState(null);
   const [hasLayout, setHasLayout] = useState(false);
   const [showInspectOverlay, setShowInspectOverlay] = useState(false);
-  const [isInspecting, setIsInspecting] = useState(false);
+  const [showDimensionsBox, setShowDimensionsBox] = useState(false);
   const [inspectOverlayFrame, setInspectOverlayFrame] = useState({
     x: 0,
     y: 0,
@@ -310,28 +311,30 @@ export function AppWrapper({ children, initialProps, fabric }) {
     devtoolsAgent,
     "RNIDE_inspect",
     (payload) => {
-      const { id, x, y, requestStack } = payload;
+      const { id, x, y, requestStack, showDimensionsBox } = payload;
 
       getInspectorDataForCoordinates(mainContainerRef, x, y, requestStack, (inspectorData) => {
         devtoolsAgent._bridge.send("RNIDE_inspectData", {
           id,
           ...inspectorData,
         });
+        setInspectOverlayFrame(inspectorData.frame);
+        setShowDimensionsBox(showDimensionsBox);
+        setShowInspectOverlay(true);
       });
     },
     [mainContainerRef]
   );
 
   useAgentListener(devtoolsAgent, "RNIDE_showInspectOverlay", (payload) => {
-    const {x, y, width, height, isInspecting} = payload;
-    setInspectOverlayFrame({x,y,height,width});
-    setIsInspecting(!!isInspecting);
+    const { x, y, width, height } = payload;
+    setInspectOverlayFrame({ x, y, height, width });
     setShowInspectOverlay(true);
   });
 
   useAgentListener(devtoolsAgent, "RNIDE_hideInspectOverlay", (payload) => {
     setShowInspectOverlay(false);
-    setIsInspecting(false);
+    setShowDimensionsBox(false);
   });
 
   useAgentListener(
@@ -401,7 +404,7 @@ export function AppWrapper({ children, initialProps, fabric }) {
             backgroundColor: "rgba(56, 172, 221, 0.85)",
           }}></View>
       )}
-      {showInspectOverlay && isInspecting && <DimensionsBox frame={inspectOverlayFrame} />}
+      {showInspectOverlay && showDimensionsBox && <DimensionsBox frame={inspectOverlayFrame} />}
     </>
   );
 }
@@ -416,97 +419,4 @@ export function createNestedAppWrapper(InnerWrapperComponent) {
     );
   }
   return WrapperComponent;
-}
-
-// This section should be imported but importing components to this file does not work for some reason 
-const BOX_HEIGHT_FRACTION = 0.033;
-const FONT_SIZE_FRACTION = 0.6;
-const BORDER_RADIUS_FRACTION = 0.15;
-const HORIZONTAL_PADDING_FRACTION = 0.3;
-
-const VERTICAL_POSITION_THRESHOLD = 0.3;
-const HORIZONTAL_POSITION_THRESHOLD = 0.5;
-
-const VERTICAL_ARROW_MARGIN = 10;
-const HORIZONTAL_ARROW_MARGIN = 20;
-
-function DimensionsBox({ frame }) {
-  const width = parseFloat(frame.width.toFixed(2));
-  const height = parseFloat(frame.height.toFixed(2));
-  const { width: screenWidth, height: screenHeight } = Dimensions.get("screen");
-
-  const boxPosition = (() => {
-    if (frame.y / screenHeight >= VERTICAL_POSITION_THRESHOLD) {
-      return "above";
-    } else if ((frame.y + frame.height) / screenHeight <= 1 - VERTICAL_POSITION_THRESHOLD) {
-      return "below";
-    } else if ((frame.x + frame.width) / screenWidth <= HORIZONTAL_POSITION_THRESHOLD) {
-      return "right";
-    } else if (frame.x / screenWidth >= 1 - HORIZONTAL_POSITION_THRESHOLD) {
-      return "left";
-    }
-    return "inside";
-  })();
-
-  const positionalProps = (() => {
-    switch (boxPosition) {
-      case "above":
-        const res = {
-          top: frame.y - VERTICAL_ARROW_MARGIN,
-          left: frame.x + frame.width / 2,
-          transform: "translate(-50%, -100%)",
-        };
-        return res;
-      case "below":
-        return {
-          top: frame.y + frame.height + VERTICAL_ARROW_MARGIN,
-          left: frame.x + frame.width / 2,
-          transform: "translate(-50%, 0%)",
-        };
-      case "right":
-        return {
-          top: frame.y + frame.height / 2,
-          left: frame.x + frame.width + HORIZONTAL_ARROW_MARGIN,
-          transform: "translate(0%, -50%)",
-        };
-      case "left":
-        return {
-          top: frame.y + frame.height / 2,
-          left: frame.x - HORIZONTAL_ARROW_MARGIN,
-          transform: "translate(-100%, -50%)",
-        };
-      default:
-        return {
-          top: frame.y + frame.height / 2,
-          left: frame.x + frame.width / 2,
-          transform: "translate(-50%, -50%)",
-        };
-    }
-  })();
-
-  const boxHeight = screenHeight * BOX_HEIGHT_FRACTION;
-
-  const cssPropertiesForDimensionsBox = {
-    height: boxHeight,
-    justifyContent: "center",
-    paddingLeft: boxHeight * HORIZONTAL_PADDING_FRACTION,
-    paddingRight: boxHeight * HORIZONTAL_PADDING_FRACTION,
-    borderRadius: boxHeight * BORDER_RADIUS_FRACTION,
-    fontSize: boxHeight * FONT_SIZE_FRACTION,
-    backgroundColor: "rgb(64, 64, 64)",
-    position: "absolute",
-    alignItems: "center",
-  };
-
-  return (
-    <View
-      style={{
-        ...cssPropertiesForDimensionsBox,
-        ...positionalProps,
-      }}>
-      <Text style={{ color: "rgb(101, 123, 131)" }}>
-        {width} × {height}
-      </Text>
-    </View>
-  );
 }
