+++ packages/vscode-extension/src/project/deviceSession.ts
@@ -288,6 +288,7 @@ export class DeviceSession implements Disposable {
     xRatio: number,
     yRatio: number,
     requestStack: boolean,
+    showDimensionsBox: boolean,
     callback: (inspectData: any) => void
   ) {
     const id = this.inspectCallID++;
@@ -298,14 +299,17 @@ export class DeviceSession implements Disposable {
       }
     };
     this.devtools?.addListener(listener);
-    this.devtools.send("RNIDE_inspect", { x: xRatio, y: yRatio, id, requestStack });
+    this.devtools.send("RNIDE_inspect", {
+      x: xRatio,
+      y: yRatio,
+      id,
+      requestStack,
+      showDimensionsBox,
+    });
   }
 
-  public showInspectOverlay(
-    frame: { x: number; y: number; width: number; height: number },
-    isInspecting?: boolean
-  ): void {
-    this.devtools.send("RNIDE_showInspectOverlay", { ...frame, isInspecting });
+  public showInspectOverlay(frame: { x: number; y: number; width: number; height: number }): void {
+    this.devtools.send("RNIDE_showInspectOverlay", frame);
   }
 
   public hideInspectOverlay() {
