+++ packages/vscode-extension/src/webview/views/PreviewView.tsx
@@ -279,7 +279,7 @@ function PreviewView() {
           onSelected={onInspectorItemSelected}
           onHover={(item) => {
             if (item.frame) {
-              project.showInspectOverlay(item.frame, isInspecting);
+              project.showInspectOverlay(item.frame);
               setInspectFrame(item.frame);
             }
           }}
